requirements.txt
Pillow
albumentations
opencv-python-headless
numpy
tqdm
concurrent-futures
torch